% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\docType{class}
\name{AbstractWrapper}
\alias{AbstractWrapper}
\title{AbstractWrapper Class}
\description{
Abstract class representing a local dataset object in a Vitessce dataset.
}
\details{
Abstract dataset object wrapper class
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{out_dir}}{The directory for processed output files.}

\item{\code{use_cache}}{If converted output files already exist, should they be used?}

\item{\code{is_remote}}{Is the data object fully remote?}

\item{\code{routes}}{A list of \code{VitessceConfigServerRoute} objects.}

\item{\code{file_def_creators}}{A list of file definition creator functions.}

\item{\code{base_dir}}{The base directory for local data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractWrapper-new}{\code{AbstractWrapper$new()}}
\item \href{#method-AbstractWrapper-convert_and_save}{\code{AbstractWrapper$convert_and_save()}}
\item \href{#method-AbstractWrapper-get_routes}{\code{AbstractWrapper$get_routes()}}
\item \href{#method-AbstractWrapper-get_file_defs}{\code{AbstractWrapper$get_file_defs()}}
\item \href{#method-AbstractWrapper-get_out_dir_route}{\code{AbstractWrapper$get_out_dir_route()}}
\item \href{#method-AbstractWrapper-get_local_dir_url}{\code{AbstractWrapper$get_local_dir_url()}}
\item \href{#method-AbstractWrapper-get_local_file_url}{\code{AbstractWrapper$get_local_file_url()}}
\item \href{#method-AbstractWrapper-get_local_dir_route}{\code{AbstractWrapper$get_local_dir_route()}}
\item \href{#method-AbstractWrapper-get_local_file_route}{\code{AbstractWrapper$get_local_file_route()}}
\item \href{#method-AbstractWrapper-get_url}{\code{AbstractWrapper$get_url()}}
\item \href{#method-AbstractWrapper-get_url_simple}{\code{AbstractWrapper$get_url_simple()}}
\item \href{#method-AbstractWrapper-get_route_str}{\code{AbstractWrapper$get_route_str()}}
\item \href{#method-AbstractWrapper-get_out_dir}{\code{AbstractWrapper$get_out_dir()}}
\item \href{#method-AbstractWrapper-auto_view_config}{\code{AbstractWrapper$auto_view_config()}}
\item \href{#method-AbstractWrapper-clone}{\code{AbstractWrapper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-new"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-new}{}}}
\subsection{Method \code{new()}}{
Create an abstract wrapper around a data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$new(out_dir = NA, use_cache = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_dir}}{The directory for processed output files.}

\item{\code{use_cache}}{If converted output files already exist, should they be used? By default, FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AbstractWrapper} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-convert_and_save"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-convert_and_save}{}}}
\subsection{Method \code{convert_and_save()}}{
Fill in the file_def_creators array.
Each function added to this list should take in a base URL and generate a Vitessce file definition.
If this wrapper is wrapping local data, then create routes and fill in the routes array.
This method is void, should not return anything.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$convert_and_save(dataset_uid, obj_i, base_dir = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{A unique identifier for this dataset.}

\item{\code{obj_i}}{Within the dataset, the index of this data wrapper object.}

\item{\code{base_dir}}{Path to a base directory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_routes"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_routes}{}}}
\subsection{Method \code{get_routes()}}{
Obtain the routes that have been created for this wrapper class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_routes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of server routes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_file_defs"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_file_defs}{}}}
\subsection{Method \code{get_file_defs()}}{
Obtain the file definitions for this wrapper class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_file_defs(base_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{A base URL to prepend to relative URLs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of server routes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_out_dir_route"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_out_dir_route}{}}}
\subsection{Method \code{get_out_dir_route()}}{
Create a web server route for this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_out_dir_route(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigServerStaticRoute} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_local_dir_url"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_local_dir_url}{}}}
\subsection{Method \code{get_local_dir_url()}}{
Construct a URL to a local directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_local_dir_url(
  base_url,
  dataset_uid,
  obj_i,
  local_dir_path,
  local_dir_uid
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{The base URL on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{local_dir_path}}{The path to the local directory.}

\item{\code{local_dir_uid}}{A unique identifier for this local directory in this dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string for the URL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_local_file_url"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_local_file_url}{}}}
\subsection{Method \code{get_local_file_url()}}{
Construct a URL to a local file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_local_file_url(
  base_url,
  dataset_uid,
  obj_i,
  local_file_path,
  local_file_uid
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{The base URL on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{local_file_path}}{The path to the local file.}

\item{\code{local_file_uid}}{A unique identifier for this local file in this dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A string for the URL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_local_dir_route"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_local_dir_route}{}}}
\subsection{Method \code{get_local_dir_route()}}{
Create a web server route for this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_local_dir_route(
  dataset_uid,
  obj_i,
  local_dir_path,
  local_dir_uid
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{local_dir_path}}{The path to the local directory.}

\item{\code{local_dir_uid}}{A unique identifier for this local directory in this dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigServerStaticRoute} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_local_file_route"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_local_file_route}{}}}
\subsection{Method \code{get_local_file_route()}}{
Create a web server route for this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_local_file_route(
  dataset_uid,
  obj_i,
  local_file_path,
  local_file_uid
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{local_file_path}}{The path to the local file.}

\item{\code{local_file_uid}}{A unique identifier for this local file in this dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigServerFileRoute} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_url}{}}}
\subsection{Method \code{get_url()}}{
Create a local web server URL for a dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_url(base_url, dataset_uid, obj_i, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{The base URL on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{...}}{Passes extra arguments to \code{get_route_str}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A URL as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_url_simple"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_url_simple}{}}}
\subsection{Method \code{get_url_simple()}}{
Construct a URL.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_url_simple(base_url, suffix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{The base URL on which the web server is serving.}

\item{\code{suffix}}{The suffix to append to the base URL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A URL as a string like {base_url}/{suffix}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_route_str"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_route_str}{}}}
\subsection{Method \code{get_route_str()}}{
Create a string representing a web server route path (the part following the base URL).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_route_str(dataset_uid, obj_i, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{...}}{Pastes extra arguments together after \code{obj_i}, separated by slash.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A path as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-get_out_dir"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-get_out_dir}{}}}
\subsection{Method \code{get_out_dir()}}{
Create a directory path to a dataset within the base output directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_out_dir(dataset_uid, obj_i, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{...}}{Passes extra arguments to \code{file.path}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A path as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-auto_view_config"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-auto_view_config}{}}}
\subsection{Method \code{auto_view_config()}}{
Automatically configure views for a particular dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$auto_view_config(vc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vc}}{A \code{VitessceConfig} instance to configure.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractWrapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
