% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_anndata.R
\docType{class}
\name{AnnDataWrapper}
\alias{AnnDataWrapper}
\title{AnnDataWrapper Class}
\description{
Class representing an AnnData object (saved to a Zarr store) in a Vitessce dataset.
}
\details{
AnnData object wrapper class
}
\keyword{internal}
\section{Super class}{
\code{\link[vitessceR:AbstractWrapper]{vitessceR::AbstractWrapper}} -> \code{AnnDataWrapper}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{adata_path}}{The object to wrap.}

\item{\code{adata_url}}{The object to wrap.}

\item{\code{local_dir_uid}}{The path to the local zarr store.}

\item{\code{obs_feature_matrix_path}}{The path to the observation-by-feature matrix within the Zarr store.}

\item{\code{feature_filter_path}}{The path to the a column of adata.var within the Zarr store.
Use this if obs_feature_matrix_path points to a subset of adata.X (relative to the full adata.var).}

\item{\code{initial_feature_filter_path}}{The path to the a column of adata.var within the Zarr store.
Use this to load a subset of the matrix at obs_feature_matrix_path initially.}

\item{\code{obs_set_paths}}{A list of paths to the observation sets within the Zarr store.}

\item{\code{obs_set_names}}{A list of names for the observation sets.}

\item{\code{obs_locations_path}}{The path to the observation locations within the Zarr store.}

\item{\code{obs_segmentations_path}}{The path to the observation segmentations within the Zarr store.}

\item{\code{obs_embedding_paths}}{A list of paths to the observation embeddings within the Zarr store.}

\item{\code{obs_embedding_names}}{A list of names for the observation embeddings.}

\item{\code{obs_embedding_dims}}{A list of dimensions for the observation embeddings.}

\item{\code{feature_labels_path}}{The path to the feature labels within the Zarr store.}

\item{\code{obs_labels_path}}{The path to the observation labels within the Zarr store.}

\item{\code{obs_labels_paths}}{A list of paths to the observation labels within the Zarr store.}

\item{\code{obs_labels_names}}{A list of names for the observation labels.}

\item{\code{coordination_values}}{A list of coordination values for the file definition.}

\item{\code{request_init}}{A list of requestInit options for the Zarr store.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnDataWrapper-new}{\code{AnnDataWrapper$new()}}
\item \href{#method-AnnDataWrapper-convert_and_save}{\code{AnnDataWrapper$convert_and_save()}}
\item \href{#method-AnnDataWrapper-make_routes}{\code{AnnDataWrapper$make_routes()}}
\item \href{#method-AnnDataWrapper-get_zarr_url}{\code{AnnDataWrapper$get_zarr_url()}}
\item \href{#method-AnnDataWrapper-make_file_def_creator}{\code{AnnDataWrapper$make_file_def_creator()}}
\item \href{#method-AnnDataWrapper-clone}{\code{AnnDataWrapper$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="auto_view_config"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-auto_view_config'><code>vitessceR::AbstractWrapper$auto_view_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_file_defs"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_file_defs'><code>vitessceR::AbstractWrapper$get_file_defs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_local_dir_route"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_local_dir_route'><code>vitessceR::AbstractWrapper$get_local_dir_route()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_local_dir_url"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_local_dir_url'><code>vitessceR::AbstractWrapper$get_local_dir_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_local_file_route"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_local_file_route'><code>vitessceR::AbstractWrapper$get_local_file_route()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_local_file_url"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_local_file_url'><code>vitessceR::AbstractWrapper$get_local_file_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_out_dir"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_out_dir'><code>vitessceR::AbstractWrapper$get_out_dir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_out_dir_route"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_out_dir_route'><code>vitessceR::AbstractWrapper$get_out_dir_route()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_route_str"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_route_str'><code>vitessceR::AbstractWrapper$get_route_str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_routes"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_routes'><code>vitessceR::AbstractWrapper$get_routes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_url"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_url'><code>vitessceR::AbstractWrapper$get_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_url_simple"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_url_simple'><code>vitessceR::AbstractWrapper$get_url_simple()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataWrapper-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataWrapper-new}{}}}
\subsection{Method \code{new()}}{
Create a wrapper around an AnnData object saved to a Zarr store.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataWrapper$new(
  adata_path = NA,
  adata_url = NA,
  obs_feature_matrix_path = NA,
  feature_filter_path = NA,
  initial_feature_filter_path = NA,
  obs_set_paths = NA,
  obs_set_names = NA,
  obs_locations_path = NA,
  obs_segmentations_path = NA,
  obs_embedding_paths = NA,
  obs_embedding_names = NA,
  obs_embedding_dims = NA,
  request_init = NA,
  feature_labels_path = NA,
  coordination_values = NA,
  obs_labels_paths = NA,
  obs_labels_names = NA,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adata_path}}{The path to a local Zarr store.}

\item{\code{adata_url}}{The URL to a remote Zarr store.}

\item{\code{obs_feature_matrix_path}}{The path to the observation-by-feature matrix within the Zarr store.}

\item{\code{feature_filter_path}}{The path to the a column of adata.var within the Zarr store. Use this if obs_feature_matrix_path points to a subset of adata.X (relative to the full adata.var).}

\item{\code{initial_feature_filter_path}}{The path to the a column of adata.var within the Zarr store. Use this to load a subset of the matrix at obs_feature_matrix_path initially.}

\item{\code{obs_set_paths}}{A list of paths to the observation sets within the Zarr store.}

\item{\code{obs_set_names}}{A list of names for the observation sets.}

\item{\code{obs_locations_path}}{The path to the observation locations within the Zarr store.}

\item{\code{obs_segmentations_path}}{The path to the observation segmentations within the Zarr store.}

\item{\code{obs_embedding_paths}}{A list of paths to the observation embeddings within the Zarr store.}

\item{\code{obs_embedding_names}}{A list of names for the observation embeddings.}

\item{\code{obs_embedding_dims}}{A list of dimensions for the observation embeddings.}

\item{\code{request_init}}{A list of requestInit options for the Zarr store.}

\item{\code{feature_labels_path}}{The path to the feature labels within the Zarr store.}

\item{\code{coordination_values}}{A list of coordination values for the file definition.}

\item{\code{obs_labels_paths}}{A list of paths to the observation labels within the Zarr store.}

\item{\code{obs_labels_names}}{A list of names for the observation labels.}

\item{\code{...}}{Parameters inherited from \code{AbstractWrapper}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{AnnDataWrapper} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataWrapper-convert_and_save"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataWrapper-convert_and_save}{}}}
\subsection{Method \code{convert_and_save()}}{
Create the JSON output files, web server routes, and file definition creators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataWrapper$convert_and_save(dataset_uid, obj_i, base_dir = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{base_dir}}{A base directory for local data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataWrapper-make_routes"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataWrapper-make_routes}{}}}
\subsection{Method \code{make_routes()}}{
Get a list of server route objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataWrapper$make_routes(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataWrapper-get_zarr_url"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataWrapper-get_zarr_url}{}}}
\subsection{Method \code{get_zarr_url()}}{
Get the URL to the Zarr store, to fill in the file URL in the file definitions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataWrapper$get_zarr_url(base_url, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{The base URL, on which the route will be served.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A URL as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataWrapper-make_file_def_creator"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataWrapper-make_file_def_creator}{}}}
\subsection{Method \code{make_file_def_creator()}}{
Make the file definition creator function for the cells data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataWrapper$make_file_def_creator(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A file definition creator function which takes a \code{base_url} parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnDataWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnDataWrapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnDataWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
