% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{Codec}
\alias{Codec}
\title{Codec Class}
\description{
Abstract class representing a compressor.
}
\details{
Abstract compressor for Zarr
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-encode}{\code{Codec$encode()}}
\item \href{#method-decode}{\code{Codec$decode()}}
\item \href{#method-get_meta}{\code{Codec$get_meta()}}
\item \href{#method-clone}{\code{Codec$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encode"></a>}}
\if{latex}{\out{\hypertarget{method-encode}{}}}
\subsection{Method \code{encode()}}{
Compress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$encode(buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The un-compressed data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
Decompress data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$decode(buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{buf}}{The compressed data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Un-compressed data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_meta"></a>}}
\if{latex}{\out{\hypertarget{method-get_meta}{}}}
\subsection{Method \code{get_meta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$get_meta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codec$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
