% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CoordinationType}
\alias{CoordinationType}
\title{Coordination Type}
\format{
An object of class \code{list} of length 32.
}
\usage{
CoordinationType
}
\description{
The \code{CoordinationType} list contains an enumeration of
valid string constant values representing coordination types
in the coordination space.
}
\keyword{datasets}
