% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DataType}
\alias{DataType}
\title{Enumeration of Data Type values}
\format{
An object of class \code{list} of length 6.
}
\usage{
DataType
}
\description{
The \code{DataType} list contains an enumeration of
valid string constant values representing data types
for dataset files.
}
\keyword{datasets}
