% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_objects.R
\name{get_giotto_obj}
\alias{get_giotto_obj}
\title{Create a mock Giotto object for tests and examples.}
\usage{
get_giotto_obj()
}
\value{
The object.
}
\description{
Create a mock Giotto object for tests and examples.
}
\examples{
obj <- get_giotto_obj()
}
\keyword{internal}
