% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_next_scope}
\alias{get_next_scope}
\title{Get next scope name}
\usage{
get_next_scope(prev_scopes)
}
\arguments{
\item{prev_scopes}{A vector of previous scope names.}
}
\value{
The new scope name which does not conflict with any previous scope name.
}
\description{
Get next scope name
}
\keyword{internal}
