% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_objects.R
\name{get_sce_obj}
\alias{get_sce_obj}
\title{Create a mock SCE object for tests and examples.}
\usage{
get_sce_obj()
}
\value{
The object.
}
\description{
Create a mock SCE object for tests and examples.
}
\examples{
obj <- get_sce_obj()
}
\keyword{internal}
