% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_objects.R
\name{get_seurat_obj}
\alias{get_seurat_obj}
\title{Create a mock Seurat object for tests and examples.}
\usage{
get_seurat_obj()
}
\value{
The object.
}
\description{
Create a mock Seurat object for tests and examples.
}
\examples{
obj <- get_seurat_obj()
}
\keyword{internal}
