% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_objects.R
\name{get_spe_obj}
\alias{get_spe_obj}
\title{Create a mock SPE object for tests and examples.}
\usage{
get_spe_obj()
}
\value{
The object.
}
\description{
Create a mock SPE object for tests and examples.
}
\examples{
obj <- get_spe_obj()
}
\keyword{internal}
