% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_zarr.R
\name{giotto_to_anndata_zarr}
\alias{giotto_to_anndata_zarr}
\title{Save a Giotto object to an AnnData-Zarr store}
\usage{
giotto_to_anndata_zarr(giotto_obj, out_path, X_slot)
}
\arguments{
\item{giotto_obj}{The object to save.}

\item{out_path}{A path to the output Zarr store.}

\item{X_slot}{The name of the slot in the Giotto object to use for adata.X}
}
\value{
TRUE if the conversion succeeds.
}
\description{
Save a Giotto object to an AnnData-Zarr store
}
\examples{
obj <- get_giotto_obj()
giotto_to_anndata_zarr(obj, "data/giotto.zarr")
}
\keyword{internal}
