% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hconcat}
\alias{hconcat}
\title{Horizontally concatenate views}
\usage{
hconcat(...)
}
\arguments{
\item{...}{A variable number of \code{VitessceConfigView}, \code{VitessceConfigViewHConcat}, or \code{VitessceConfigViewVConcat} objects.}
}
\value{
A \code{VitessceConfigViewHConcat} object.
}
\description{
A helper function to construct a new \code{VitessceConfigViewHConcat} object based on multiple views.
}
