% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_na}
\alias{is_na}
\title{Check if a value, potentially a vector, is NA}
\usage{
is_na(val)
}
\arguments{
\item{val}{The value to check}
}
\value{
Whether the value is NA
}
\description{
Check if a value, potentially a vector, is NA
}
\keyword{internal}
