% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{json_to_raw}
\alias{json_to_raw}
\title{Convert a JSON-like list to a raw type.}
\usage{
json_to_raw(json_as_list)
}
\arguments{
\item{json_as_list}{An R list to be converted to JSON.}
}
\value{
The raw value.
}
\description{
Convert a JSON-like list to a raw type.
}
