% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcodecs.R
\docType{class}
\name{LZ4}
\alias{LZ4}
\title{LZ4 Class}
\description{
Class representing a LZ4 compressor
}
\details{
LZ4 compressor for Zarr
}
\keyword{internal}
\section{Super class}{
\code{\link[vitessce:Codec]{vitessce::Codec}} -> \code{LZ4}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{acceleration}}{The compression level.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LZ4$new()}}
\item \href{#method-encode}{\code{LZ4$encode()}}
\item \href{#method-decode}{\code{LZ4$decode()}}
\item \href{#method-get_meta}{\code{LZ4$get_meta()}}
\item \href{#method-clone}{\code{LZ4$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new LZ4 compressor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LZ4$new(acceleration = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{acceleration}}{The compression level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{LZ4} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-encode"></a>}}
\if{latex}{\out{\hypertarget{method-encode}{}}}
\subsection{Method \code{encode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LZ4$encode(buf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LZ4$decode(buf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_meta"></a>}}
\if{latex}{\out{\hypertarget{method-get_meta}{}}}
\subsection{Method \code{get_meta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LZ4$get_meta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LZ4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
