% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_numpy_friendly}
\alias{make_numpy_friendly}
\title{Make a matrix compatible with NumPy.}
\usage{
make_numpy_friendly(x, transpose = TRUE)
}
\arguments{
\item{x}{The matrix}

\item{transpose}{Should the matrix be transposed? By default, TRUE.}
}
\value{
The matrix.
}
\description{
Reference: https://github.com/theislab/zellkonverter/blob/master/R/SCE2AnnData.R#L237
}
\keyword{internal}
