% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{matrix_to_zarr}
\alias{matrix_to_zarr}
\title{Write an R matrix to a Zarr store (one chunk, no compression).}
\usage{
matrix_to_zarr(matrix, rows, cols, store, compressor = NA)
}
\arguments{
\item{matrix}{The matrix as an R matrix.}

\item{rows}{A vector of row names.}

\item{cols}{A vector of column names.}

\item{store}{The Zarr store.}
}
\description{
Write an R matrix to a Zarr store (one chunk, no compression).
}
