% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_files}
\alias{merge_files}
\title{Merge file chunks}
\usage{
merge_files(chunk_paths, out_path)
}
\arguments{
\item{chunk_paths}{An ordered list of paths to each chunk to be merged.}

\item{out_path}{Where to store the merged file.}
}
\value{
Nothing
}
\description{
Merge file chunks
}
\keyword{internal}
