% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_js}
\alias{merge_js}
\title{Merge JS file chunks}
\usage{
merge_js(force = FALSE)
}
\arguments{
\item{force}{Should the merging be done even if the merged file already exists? By default, FALSE.}
}
\value{
Success.
}
\description{
Merge JS file chunks
}
\keyword{internal}
