% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_wrappers.R
\docType{class}
\name{MyCustomSeuratWrapper}
\alias{MyCustomSeuratWrapper}
\title{MyCustomSeuratWrapper Class}
\description{
Subclass of SeuratWrapper.
}
\details{
Subclass of SeuratWrapper to deal with cell sets.
}
\keyword{internal}
\section{Super classes}{
\code{\link[vitessce:AbstractWrapper]{vitessce::AbstractWrapper}} -> \code{\link[vitessce:SeuratWrapper]{vitessce::SeuratWrapper}} -> \code{MyCustomSeuratWrapper}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cell_sets}}{The cell sets.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MyCustomSeuratWrapper$new()}}
\item \href{#method-create_cell_sets_list}{\code{MyCustomSeuratWrapper$create_cell_sets_list()}}
\item \href{#method-clone}{\code{MyCustomSeuratWrapper$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="auto_view_config">}\href{../../vitessce/html/AbstractWrapper.html#method-auto_view_config}{\code{vitessce::AbstractWrapper$auto_view_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_file_defs">}\href{../../vitessce/html/AbstractWrapper.html#method-get_file_defs}{\code{vitessce::AbstractWrapper$get_file_defs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_out_dir">}\href{../../vitessce/html/AbstractWrapper.html#method-get_out_dir}{\code{vitessce::AbstractWrapper$get_out_dir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_out_dir_route">}\href{../../vitessce/html/AbstractWrapper.html#method-get_out_dir_route}{\code{vitessce::AbstractWrapper$get_out_dir_route()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_route_str">}\href{../../vitessce/html/AbstractWrapper.html#method-get_route_str}{\code{vitessce::AbstractWrapper$get_route_str()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_routes">}\href{../../vitessce/html/AbstractWrapper.html#method-get_routes}{\code{vitessce::AbstractWrapper$get_routes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_url">}\href{../../vitessce/html/AbstractWrapper.html#method-get_url}{\code{vitessce::AbstractWrapper$get_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="SeuratWrapper" data-id="convert_and_save">}\href{../../vitessce/html/SeuratWrapper.html#method-convert_and_save}{\code{vitessce::SeuratWrapper$convert_and_save()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="SeuratWrapper" data-id="create_cells_list">}\href{../../vitessce/html/SeuratWrapper.html#method-create_cells_list}{\code{vitessce::SeuratWrapper$create_cells_list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="SeuratWrapper" data-id="create_expression_matrix_list">}\href{../../vitessce/html/SeuratWrapper.html#method-create_expression_matrix_list}{\code{vitessce::SeuratWrapper$create_expression_matrix_list()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="SeuratWrapper" data-id="make_cell_sets_file_def_creator">}\href{../../vitessce/html/SeuratWrapper.html#method-make_cell_sets_file_def_creator}{\code{vitessce::SeuratWrapper$make_cell_sets_file_def_creator()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="SeuratWrapper" data-id="make_cells_file_def_creator">}\href{../../vitessce/html/SeuratWrapper.html#method-make_cells_file_def_creator}{\code{vitessce::SeuratWrapper$make_cells_file_def_creator()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="SeuratWrapper" data-id="make_expression_matrix_file_def_creator">}\href{../../vitessce/html/SeuratWrapper.html#method-make_expression_matrix_file_def_creator}{\code{vitessce::SeuratWrapper$make_expression_matrix_file_def_creator()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a wrapper around a Seurat object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyCustomSeuratWrapper$new(obj, cell_sets, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The object to wrap.}

\item{\code{...}}{Parameters inherited from \code{AbstractWrapper}.}

\item{\code{assay}}{The assay name under the assays part of the Seurat object.}

\item{\code{cell_set_meta_names}}{An optional list of keys in the object's meta.data
list to use for creating cell sets.}

\item{\code{cell_set_meta_score_mappings}}{If cell_set_meta_names is provided, this list can
also be provided to map between meta.data keys for set annotations
and keys for annotation scores.}

\item{\code{cell_set_meta_name_mappings}}{If cell_set_meta_names is provided, this list can
also be provided to map between meta.data keys and new names to replace
the keys in the interface.}

\item{\code{num_genes}}{The number of genes to include in the expression matrix. Note: this parameter is temporary and should be changed to something more useful such as number of most variable genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SeuratWrapper} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_cell_sets_list"></a>}}
\if{latex}{\out{\hypertarget{method-create_cell_sets_list}{}}}
\subsection{Method \code{create_cell_sets_list()}}{
Create a list representing the cluster assignments in the Seurat object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyCustomSeuratWrapper$create_cell_sets_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be converted to JSON.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MyCustomSeuratWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
