% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_images.R
\docType{class}
\name{OmeTiffWrapper}
\alias{OmeTiffWrapper}
\title{OmeTiffWrapper Class}
\description{
Class representing an OME-TIFF file in a Vitessce dataset.
}
\details{
OME-TIFF object wrapper class
}
\keyword{internal}
\section{Super class}{
\code{\link[vitessceR:AbstractWrapper]{vitessceR::AbstractWrapper}} -> \code{OmeTiffWrapper}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{img_path}}{A local filepath to an OME-TIFF file.}

\item{\code{img_url}}{A remote URL of an OME-TIFF file.}

\item{\code{name}}{The display name for this OME-TIFF within Vitessce.}

\item{\code{transformation_matrix}}{A column-major ordered matrix for transforming
this image (see http://www.opengl-tutorial.org/beginners-tutorials/tutorial-3-matrices/#homogeneous-coordinates for more information).}

\item{\code{is_bitmask}}{Whether or not this image is a bitmask.}

\item{\code{is_remote}}{Whether or not this image is remote.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OmeTiffWrapper-new}{\code{OmeTiffWrapper$new()}}
\item \href{#method-OmeTiffWrapper-convert_and_save}{\code{OmeTiffWrapper$convert_and_save()}}
\item \href{#method-OmeTiffWrapper-make_raster_routes}{\code{OmeTiffWrapper$make_raster_routes()}}
\item \href{#method-OmeTiffWrapper-make_image_def}{\code{OmeTiffWrapper$make_image_def()}}
\item \href{#method-OmeTiffWrapper-make_raster_file_def_creator}{\code{OmeTiffWrapper$make_raster_file_def_creator()}}
\item \href{#method-OmeTiffWrapper-create_image_json}{\code{OmeTiffWrapper$create_image_json()}}
\item \href{#method-OmeTiffWrapper-get_img_url}{\code{OmeTiffWrapper$get_img_url()}}
\item \href{#method-OmeTiffWrapper-clone}{\code{OmeTiffWrapper$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="auto_view_config"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-auto_view_config'><code>vitessceR::AbstractWrapper$auto_view_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_file_defs"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_file_defs'><code>vitessceR::AbstractWrapper$get_file_defs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_local_dir_route"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_local_dir_route'><code>vitessceR::AbstractWrapper$get_local_dir_route()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_local_dir_url"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_local_dir_url'><code>vitessceR::AbstractWrapper$get_local_dir_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_out_dir"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_out_dir'><code>vitessceR::AbstractWrapper$get_out_dir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_out_dir_route"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_out_dir_route'><code>vitessceR::AbstractWrapper$get_out_dir_route()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_route_str"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_route_str'><code>vitessceR::AbstractWrapper$get_route_str()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_routes"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_routes'><code>vitessceR::AbstractWrapper$get_routes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_url"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_url'><code>vitessceR::AbstractWrapper$get_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="vitessceR" data-topic="AbstractWrapper" data-id="get_url_simple"><a href='../../vitessceR/html/AbstractWrapper.html#method-AbstractWrapper-get_url_simple'><code>vitessceR::AbstractWrapper$get_url_simple()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-new"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-new}{}}}
\subsection{Method \code{new()}}{
Create a wrapper around multiple image objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$new(
  img_path = NA,
  img_url = NA,
  name = "",
  transformation_matrix = NA,
  is_bitmask = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{img_path}}{A local filepath to an OME-TIFF file.}

\item{\code{img_url}}{A remote URL of an OME-TIFF file.}

\item{\code{name}}{The display name for this OME-TIFF within Vitessce.}

\item{\code{transformation_matrix}}{A column-major ordered matrix for transforming
this image (see http://www.opengl-tutorial.org/beginners-tutorials/tutorial-3-matrices/#homogeneous-coordinates for more information).}

\item{\code{is_bitmask}}{Whether or not this image is a bitmask.}

\item{\code{...}}{Parameters inherited from \code{AbstractWrapper}.}

\item{\code{use_physical_size_scaling}}{Whether or not to scale the image based on the physical size metadata stored in the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OmeTiffWrapper} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-convert_and_save"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-convert_and_save}{}}}
\subsection{Method \code{convert_and_save()}}{
Create the web server routes and file definition creators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$convert_and_save(dataset_uid, obj_i, base_dir = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-make_raster_routes"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-make_raster_routes}{}}}
\subsection{Method \code{make_raster_routes()}}{
Create a list representing the server routes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$make_raster_routes(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of server route objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-make_image_def"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-make_image_def}{}}}
\subsection{Method \code{make_image_def()}}{
Create an object representing a single image in a raster.json list of images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$make_image_def(dataset_uid, obj_i, base_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{base_url}}{The base URL for the server.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list that can be converted to JSON.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-make_raster_file_def_creator"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-make_raster_file_def_creator}{}}}
\subsection{Method \code{make_raster_file_def_creator()}}{
Make the file definition creator function for the raster data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$make_raster_file_def_creator(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A file definition creator function which takes a \code{base_url} parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-create_image_json"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-create_image_json}{}}}
\subsection{Method \code{create_image_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$create_image_json(img_url, offsets_url = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-get_img_url"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-get_img_url}{}}}
\subsection{Method \code{get_img_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$get_img_url(base_url = "", dataset_uid = "", obj_i = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OmeTiffWrapper-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OmeTiffWrapper-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OmeTiffWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
