% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\docType{class}
\name{AbstractWrapper}
\alias{AbstractWrapper}
\title{AbstractWrapper Class}
\description{
Abstract class representing a local dataset object in a Vitessce dataset.
}
\details{
Abstract dataset object wrapper class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-create_response_json}{\code{AbstractWrapper$create_response_json()}}
\item \href{#method-get_cells}{\code{AbstractWrapper$get_cells()}}
\item \href{#method-get_cell_sets}{\code{AbstractWrapper$get_cell_sets()}}
\item \href{#method-get_raster}{\code{AbstractWrapper$get_raster()}}
\item \href{#method-get_molecules}{\code{AbstractWrapper$get_molecules()}}
\item \href{#method-get_neighborhoods}{\code{AbstractWrapper$get_neighborhoods()}}
\item \href{#method-get_expression_matrix}{\code{AbstractWrapper$get_expression_matrix()}}
\item \href{#method-get_data}{\code{AbstractWrapper$get_data()}}
\item \href{#method-get_url}{\code{AbstractWrapper$get_url()}}
\item \href{#method-get_route}{\code{AbstractWrapper$get_route()}}
\item \href{#method-clone}{\code{AbstractWrapper$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_response_json"></a>}}
\if{latex}{\out{\hypertarget{method-create_response_json}{}}}
\subsection{Method \code{create_response_json()}}{
Create a response function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$create_response_json(data_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_list}}{The data to serve, as a list. Will be converted to JSON.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new response function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cells"></a>}}
\if{latex}{\out{\hypertarget{method-get_cells}{}}}
\subsection{Method \code{get_cells()}}{
Get the routes and file definitions for the cells data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_cells(port, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or a list of \code{routes} and \code{file_defs} lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cell_sets"></a>}}
\if{latex}{\out{\hypertarget{method-get_cell_sets}{}}}
\subsection{Method \code{get_cell_sets()}}{
Get the routes and file definitions for the cell sets data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_cell_sets(port, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or a list of \code{routes} and \code{file_defs} lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_raster"></a>}}
\if{latex}{\out{\hypertarget{method-get_raster}{}}}
\subsection{Method \code{get_raster()}}{
Get the routes and file definitions for the raster type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_raster(port, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or a list of \code{routes} and \code{file_defs} lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_molecules"></a>}}
\if{latex}{\out{\hypertarget{method-get_molecules}{}}}
\subsection{Method \code{get_molecules()}}{
Get the routes and file definitions for the molecules data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_molecules(port, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or a list of \code{routes} and \code{file_defs} lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_neighborhoods"></a>}}
\if{latex}{\out{\hypertarget{method-get_neighborhoods}{}}}
\subsection{Method \code{get_neighborhoods()}}{
Get the routes and file definitions for the neighborhoods data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_neighborhoods(port, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or a list of \code{routes} and \code{file_defs} lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_expression_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-get_expression_matrix}{}}}
\subsection{Method \code{get_expression_matrix()}}{
Get the routes and file definitions for the expression matrix data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_expression_matrix(port, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or a list of \code{routes} and \code{file_defs} lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the routes and file definitions for a data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_data(data_type, port, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_type}}{A data type.}

\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or a list of \code{routes} and \code{file_defs} lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_url}{}}}
\subsection{Method \code{get_url()}}{
Create a local web server URL for a dataset object data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_url(port, dataset_uid, obj_i, suffix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The port on which the web server is serving.}

\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{suffix}}{A suffix for the URL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A URL as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_route"></a>}}
\if{latex}{\out{\hypertarget{method-get_route}{}}}
\subsection{Method \code{get_route()}}{
Create a web server route path for a dataset object data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$get_route(dataset_uid, obj_i, suffix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}

\item{\code{suffix}}{A suffix for the URL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A path as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
