% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{class}
\name{VitessceConfig}
\alias{VitessceConfig}
\title{VitessceConfig Class}
\description{
Class representing a Vitessce view config.
}
\details{
VitessceConfig
}
\examples{

## ------------------------------------------------
## Method `VitessceConfig$add_dataset`
## ------------------------------------------------

vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")

## ------------------------------------------------
## Method `VitessceConfig$add_view`
## ------------------------------------------------

vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
spatial <- vc$add_view(ds, "spatial")

## ------------------------------------------------
## Method `VitessceConfig$add_coordination`
## ------------------------------------------------

vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
spatial <- vc$add_view(ds, "spatial")
c_scopes <- vc$add_coordination(c("spatialZoom", "spatialTargetX"))
spatial$use_coordination(c_scopes)

## ------------------------------------------------
## Method `VitessceConfig$layout`
## ------------------------------------------------

vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
spatial <- vc$add_view(ds, "spatial")
scatterplot <- vc$add_view(ds, "scatterplot")

## ------------------------------------------------
## Method `VitessceConfig$to_list`
## ------------------------------------------------

vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
vc_list <- vc$to_list()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VitessceConfig$new()}}
\item \href{#method-add_dataset}{\code{VitessceConfig$add_dataset()}}
\item \href{#method-add_view}{\code{VitessceConfig$add_view()}}
\item \href{#method-add_coordination}{\code{VitessceConfig$add_coordination()}}
\item \href{#method-layout}{\code{VitessceConfig$layout()}}
\item \href{#method-to_list}{\code{VitessceConfig$to_list()}}
\item \href{#method-clone}{\code{VitessceConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new config object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfig$new(name = NA, description = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A name for the config.}

\item{\code{description}}{A description for the config.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfig} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-add_dataset}{}}}
\subsection{Method \code{add_dataset()}}{
Add a dataset to the config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfig$add_dataset(name, uid = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A name for the dataset.}

\item{\code{uid}}{A unique ID for the dataset. Optional. Created automatically if not provided.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigDataset} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_view"></a>}}
\if{latex}{\out{\hypertarget{method-add_view}{}}}
\subsection{Method \code{add_view()}}{
Add a view to the config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfig$add_view(
  dataset,
  component,
  x = NA,
  y = NA,
  w = NA,
  h = NA,
  mapping = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{The dataset object to associate with this view.}

\item{\code{component}}{The name of a component to render in this view.}

\item{\code{x}}{The x-coordinate of the view in the layout.}

\item{\code{y}}{The y-coordinate of the view in the layout.}

\item{\code{w}}{The width of the view in the layout.}

\item{\code{h}}{The height of the view in the layout.}

\item{\code{mapping}}{An optional convenience parameter for setting the \code{CoordinationType$EMBEDDING_TYPE} coordination value if the component is\code{Component$SCATTERPLOT}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigView} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
spatial <- vc$add_view(ds, "spatial")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_coordination"></a>}}
\if{latex}{\out{\hypertarget{method-add_coordination}{}}}
\subsection{Method \code{add_coordination()}}{
Add a coordination scope to the config, for one or more coordination types.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfig$add_coordination(c_types)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{c_types}}{The coordination types for which to create new coordination scopes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of new \code{VitessceConfigCoordinationScope} objects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
spatial <- vc$add_view(ds, "spatial")
c_scopes <- vc$add_coordination(c("spatialZoom", "spatialTargetX"))
spatial$use_coordination(c_scopes)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-layout"></a>}}
\if{latex}{\out{\hypertarget{method-layout}{}}}
\subsection{Method \code{layout()}}{
Define the layout of views.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfig$layout(view_concat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{view_concat}}{A concatenation of views.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
spatial <- vc$add_view(ds, "spatial")
scatterplot <- vc$add_view(ds, "scatterplot")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the config to an R list. Helpful when converting the config to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfig$to_list(on_obj = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{on_obj}}{An optional function to call upon encountering a dataset object. Used internally by the htmlwidget when rendering local objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list} that can be serialized to JSON using \code{rjson}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{vc <- VitessceConfig$new("My config")
ds <- vc$add_dataset("My dataset")
vc_list <- vc$to_list()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
