% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{class}
\name{VitessceConfigDataset}
\alias{VitessceConfigDataset}
\title{VitessceConfigDataset Class}
\description{
Class representing a dataset in a Vitessce view config.
}
\details{
Dataset in a VitessceConfig
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset}}{The dataset as a list.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VitessceConfigDataset$new()}}
\item \href{#method-add_file}{\code{VitessceConfigDataset$add_file()}}
\item \href{#method-add_object}{\code{VitessceConfigDataset$add_object()}}
\item \href{#method-to_list}{\code{VitessceConfigDataset$to_list()}}
\item \href{#method-clone}{\code{VitessceConfigDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$new(uid, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uid}}{A unique identifier string for the dataset.}

\item{\code{name}}{A name for the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigDataset} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_file"></a>}}
\if{latex}{\out{\hypertarget{method-add_file}{}}}
\subsection{Method \code{add_file()}}{
Add a file to this dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$add_file(url, data_type, file_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The URL to the file.}

\item{\code{data_type}}{The data type for the file.}

\item{\code{file_type}}{The file type for the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self, to allow chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_object"></a>}}
\if{latex}{\out{\hypertarget{method-add_object}{}}}
\subsection{Method \code{add_object()}}{
Add a data object to this dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$add_object(obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data object to add.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self, to allow chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the dataset to an R list. Helpful when converting the config to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$to_list(on_obj = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{on_obj}}{An optional function to call upon encountering a dataset object. Used internally by the htmlwidget when rendering local objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list} that can be serialized to JSON using \code{rjson}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
