% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{VitessceConfigServer}
\alias{VitessceConfigServer}
\title{R6 Class representing a local web server to serve dataset objects.}
\description{
R6 Class representing a local web server to serve dataset objects.

R6 Class representing a local web server to serve dataset objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VitessceConfigServer$new()}}
\item \href{#method-on_obj}{\code{VitessceConfigServer$on_obj()}}
\item \href{#method-run}{\code{VitessceConfigServer$run()}}
\item \href{#method-clone}{\code{VitessceConfigServer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new server wrapper object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServer$new(port)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{The server port.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigServer} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-on_obj"></a>}}
\if{latex}{\out{\hypertarget{method-on_obj}{}}}
\subsection{Method \code{on_obj()}}{
Callback for a dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServer$on_obj(obj, dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The dataset object to serve.}

\item{\code{dataset_uid}}{The dataset UID.}

\item{\code{obj_i}}{The object index within the dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run the local server on the specified port.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServer$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
