% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_zarr.R
\name{sce_to_anndata_zarr}
\alias{sce_to_anndata_zarr}
\title{Save a SingleCellExperiment to an AnnData-Zarr store.}
\usage{
sce_to_anndata_zarr(sce_obj, out_path)
}
\arguments{
\item{sce_obj}{The object to save.}

\item{out_path}{A path to the output Zarr store.}
}
\description{
Save a SingleCellExperiment to an AnnData-Zarr store.
}
\keyword{internal}
