% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_zarr.R
\name{seurat_to_anndata_zarr}
\alias{seurat_to_anndata_zarr}
\title{Save a Seurat object to an AnnData-Zarr store.}
\usage{
seurat_to_anndata_zarr(seurat_obj, out_path, assay)
}
\arguments{
\item{seurat_obj}{The object to save.}

\item{out_path}{A path to the output Zarr store.}

\item{assay}{The name of the assay to save.}
}
\value{
TRUE if the conversion succeeds.
}
\description{
Save a Seurat object to an AnnData-Zarr store.
}
\keyword{internal}
