% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_zarr.R
\name{spe_to_anndata_zarr}
\alias{spe_to_anndata_zarr}
\title{Save a SpatialExperiment to an AnnData-Zarr store.}
\usage{
spe_to_anndata_zarr(spe_obj, out_path)
}
\arguments{
\item{spe_obj}{The object to save.}

\item{out_path}{A path to the output Zarr store.}
}
\description{
Save a SpatialExperiment to an AnnData-Zarr store.
}
\keyword{internal}
