% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_zarr.R
\name{spe_to_ome_zarr}
\alias{spe_to_ome_zarr}
\title{Save an image in a SpatialExperiment to an OME-Zarr store}
\usage{
spe_to_ome_zarr(spe_obj, sample_id, image_id, out_path)
}
\arguments{
\item{spe_obj}{The object containing the image.}

\item{sample_id}{The sample_id in the imgData data frame.}

\item{image_id}{The image_id in the imgData data frame.}

\item{out_path}{A path to the output Zarr store.}
}
\value{
TRUE if the conversion succeeds.
}
\description{
Save an image in a SpatialExperiment to an OME-Zarr store
}
\examples{
obj <- get_spe_obj()
spe_to_ome_zarr(obj, "sample01", "image01", "data/spe_image.zarr")
}
\keyword{internal}
