% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vitessce_config}
\alias{vitessce_config}
\title{Create a new Vitessce Config}
\usage{
vitessce_config(name = "", description = "")
}
\arguments{
\item{name}{A name for the view config.}

\item{description}{A description for the view config.}
}
\description{
A helper function to construct a new \code{VitessceConfig} object.
}
\examples{
vc <- vitessce_config("My config")
# On the returned object, access the methods of the VitessceConfig class.
ds <- vc$add_dataset("My dataset")
}
