% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vitessce_config_from_list}
\alias{vitessce_config_from_list}
\title{Create a new Vitessce Config from a list}
\usage{
vitessce_config_from_list(config)
}
\arguments{
\item{config}{A list containing a valid config.}
}
\value{
A \code{VitessceConfig} object reflecting the list-based configuration values.
}
\description{
A helper function to construct a new \code{VitessceConfig} object based on an existing config in a list format.
}
