% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vitessce_config_from_object}
\alias{vitessce_config_from_object}
\title{Create a new Vitessce Config from a data object}
\usage{
vitessce_config_from_object(obj, name = NA, description = NA)
}
\arguments{
\item{obj}{An object from which to construct a config. Can be a SingleCellExperiment or Seurat object.}

\item{name}{A name for the view config.}

\item{description}{A description for the view config.}
}
\value{
A \code{VitessceConfig} object containing the object as a member of the datasets list, with some automatically-configured views.
}
\description{
A helper function to construct a new \code{VitessceConfig} object based on an object containing single-cell or imaging data.
}
