% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{VitessceConfig$from_list}
\alias{VitessceConfig$from_list}
\title{Create a new Vitessce Config from a list

A helper function to construct a new \code{VitessceConfig} object based on an existing config in a list format.}
\arguments{
\item{config}{A list containing a valid config.}
}
\value{
A \code{VitessceConfig} object reflecting the list-based configuration values.
}
\description{
Create a new Vitessce Config from a list

A helper function to construct a new \code{VitessceConfig} object based on an existing config in a list format.
}
