% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{VitessceConfig$from_object}
\alias{VitessceConfig$from_object}
\title{Create a new Vitessce Config from a data object

A helper function to construct a new \code{VitessceConfig} object based on an object containing single-cell or imaging data.}
\arguments{
\item{obj}{An object from which to construct a config. Can be a SingleCellExperiment or Seurat object.}

\item{name}{A name for the view config.}

\item{description}{A description for the view config.}
}
\value{
A \code{VitessceConfig} object containing the object as a member of the datasets list, with some automatically-configured views.
}
\description{
Create a new Vitessce Config from a data object

A helper function to construct a new \code{VitessceConfig} object based on an object containing single-cell or imaging data.
}
