% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vitessceConfig}
\alias{vitessceConfig}
\title{Vitessce Config}
\usage{
vitessceConfig(name = "", description = "")
}
\arguments{
\item{name}{A name for the view config.}

\item{description}{A description for the view config.}
}
\description{
Vitessce Config
}
