% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{class}
\name{VitessceConfigDataset}
\alias{VitessceConfigDataset}
\title{VitessceConfigDataset Class}
\description{
Class representing a dataset in a Vitessce view config.
}
\details{
Dataset in a VitessceConfig
}
\examples{

## ------------------------------------------------
## Method `VitessceConfigDataset$add_file`
## ------------------------------------------------

base_url <- "http://localhost:8000/"
vc <- VitessceConfig$new("My config")
dataset <- vc$add_dataset("My dataset")$add_file(
  url = paste0(base_url, "cells.json"),
  data_type = DataType$CELLS,
  file_type = FileType$CELLS_JSON
)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataset}}{The dataset as a list.}

\item{\code{base_dir}}{The base directory for the config.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VitessceConfigDataset-new}{\code{VitessceConfigDataset$new()}}
\item \href{#method-VitessceConfigDataset-get_name}{\code{VitessceConfigDataset$get_name()}}
\item \href{#method-VitessceConfigDataset-get_uid}{\code{VitessceConfigDataset$get_uid()}}
\item \href{#method-VitessceConfigDataset-add_file}{\code{VitessceConfigDataset$add_file()}}
\item \href{#method-VitessceConfigDataset-add_object}{\code{VitessceConfigDataset$add_object()}}
\item \href{#method-VitessceConfigDataset-get_routes}{\code{VitessceConfigDataset$get_routes()}}
\item \href{#method-VitessceConfigDataset-to_list}{\code{VitessceConfigDataset$to_list()}}
\item \href{#method-VitessceConfigDataset-clone}{\code{VitessceConfigDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$new(uid, name, base_dir = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uid}}{A unique identifier string for the dataset.}

\item{\code{name}}{A name for the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigDataset} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-get_name}{}}}
\subsection{Method \code{get_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-get_uid"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-get_uid}{}}}
\subsection{Method \code{get_uid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$get_uid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-add_file"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-add_file}{}}}
\subsection{Method \code{add_file()}}{
Add a file to this dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$add_file(
  url = NA,
  file_type = NA,
  options = NA,
  coordination_values = NA,
  data_type = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The URL to the file.}

\item{\code{file_type}}{The file type for the file.}

\item{\code{options}}{Optional. An options list for the file.}

\item{\code{data_type}}{The data type for the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self, to allow chaining.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{base_url <- "http://localhost:8000/"
vc <- VitessceConfig$new("My config")
dataset <- vc$add_dataset("My dataset")$add_file(
  url = paste0(base_url, "cells.json"),
  data_type = DataType$CELLS,
  file_type = FileType$CELLS_JSON
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-add_object"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-add_object}{}}}
\subsection{Method \code{add_object()}}{
Add a data object to this dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$add_object(obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The data object to add.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self, to allow chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-get_routes"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-get_routes}{}}}
\subsection{Method \code{get_routes()}}{
Get a list of web server route objects corresponding to any local files which will need to be served.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$get_routes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{list} of \code{VitessceConfigServerStaticRoute}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the dataset to an R list. Helpful when converting the config to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$to_list(base_url = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{The base URL to prepend to web server file paths.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list} that can be serialized to JSON using \code{rjson}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
