% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{class}
\name{VitessceConfigDatasetFile}
\alias{VitessceConfigDatasetFile}
\title{VitessceConfigDatasetFile Class}
\description{
Class representing a file in a dataset in a Vitessce view config.
}
\details{
File in a dataset in a VitessceConfig
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VitessceConfigDatasetFile-new}{\code{VitessceConfigDatasetFile$new()}}
\item \href{#method-VitessceConfigDatasetFile-to_list}{\code{VitessceConfigDatasetFile$to_list()}}
\item \href{#method-VitessceConfigDatasetFile-clone}{\code{VitessceConfigDatasetFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDatasetFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDatasetFile-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataset file object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDatasetFile$new(
  url = NA,
  data_type = NA,
  file_type = NA,
  options = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{A URL for the file.}

\item{\code{data_type}}{A data type for the file.}

\item{\code{file_type}}{A file type for the file.}

\item{\code{options}}{A list of options to include for the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigDatasetFile} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDatasetFile-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDatasetFile-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to an R list. Helpful when converting the config to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDatasetFile$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{list} that can be serialized to JSON using \code{rjson}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VitessceConfigDatasetFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VitessceConfigDatasetFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigDatasetFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
