% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vitessceConfigFromList}
\alias{vitessceConfigFromList}
\title{Vitessce Config from list}
\usage{
vitessceConfigFromList(config)
}
\arguments{
\item{config}{A list containing a valid config.}
}
\description{
Vitessce Config from list
}
