% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vitessceConfigFromObject}
\alias{vitessceConfigFromObject}
\title{Vitessce Config from object}
\usage{
vitessceConfigFromObject(obj, name = NA, description = NA)
}
\arguments{
\item{obj}{An object from which to construct a config. Can be a SingleCellExperiment or Seurat object.}

\item{name}{A name for the view config.}

\item{description}{A description for the view config.}
}
\description{
Vitessce Config from object
}
