% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{VitessceConfigServerCallbackRoute}
\alias{VitessceConfigServerCallbackRoute}
\title{Class representing a local web server route: path + callback.}
\description{
Class representing a local web server route: path + callback.

Class representing a local web server route: path + callback.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{The path on which the web server should respond to requests using this callback.}

\item{\code{callback}}{The response callback.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VitessceConfigServerCallbackRoute$new()}}
\item \href{#method-clone}{\code{VitessceConfigServerCallbackRoute$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new server route wrapper object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServerCallbackRoute$new(path, callback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The route path.}

\item{\code{callback}}{The response callback.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigServerCallbackRoute} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServerCallbackRoute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
