% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{class}
\name{VitessceConfigView}
\alias{VitessceConfigView}
\title{VitessceConfigView Class}
\description{
Class representing a view in a Vitessce view config.
}
\details{
View in a VitessceConfig
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VitessceConfigView$new()}}
\item \href{#method-use_coordination}{\code{VitessceConfigView$use_coordination()}}
\item \href{#method-set_xywh}{\code{VitessceConfigView$set_xywh()}}
\item \href{#method-set_props}{\code{VitessceConfigView$set_props()}}
\item \href{#method-to_list}{\code{VitessceConfigView$to_list()}}
\item \href{#method-clone}{\code{VitessceConfigView$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new view object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigView$new(component, coordination_scopes, x, y, w, h)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component}}{A component name.}

\item{\code{coordination_scopes}}{A list of the coordination scope mappings for each coordination type.}

\item{\code{x}}{The x-coordinate of the view in the layout.}

\item{\code{y}}{The y-coordinate of the view in the layout.}

\item{\code{w}}{The width of the view in the layout.}

\item{\code{h}}{The height of the view in the layout.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigView} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-use_coordination"></a>}}
\if{latex}{\out{\hypertarget{method-use_coordination}{}}}
\subsection{Method \code{use_coordination()}}{
Link this view to existing coordination scope(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigView$use_coordination(c_scopes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{c_scopes}}{A list of \code{VitessceConfigCoordinationScope} objects to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self, to allow chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_xywh"></a>}}
\if{latex}{\out{\hypertarget{method-set_xywh}{}}}
\subsection{Method \code{set_xywh()}}{
Set the dimensions of the view.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigView$set_xywh(x, y, w, h)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x-coordinate of the view in the layout.}

\item{\code{y}}{The y-coordinate of the view in the layout.}

\item{\code{w}}{The width of the view in the layout.}

\item{\code{h}}{The height of the view in the layout.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, to allow chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_props"></a>}}
\if{latex}{\out{\hypertarget{method-set_props}{}}}
\subsection{Method \code{set_props()}}{
Set additional props for this view.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigView$set_props(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any named props to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, to allow chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to an R list. Helpful when converting the config to JSON.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigView$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{list} that can be serialized to JSON using \code{rjson}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigView$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
