% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{vitessceWidget}
\alias{vitessceWidget}
\title{Vitessce Widget}
\usage{
vitessceWidget(
  config = NULL,
  theme = "dark",
  width = NULL,
  height = NULL,
  port = 8000,
  elementId = NULL
)
}
\arguments{
\item{config}{A view config as a \code{VitessceConfig} object.}

\item{theme}{The theme of the widget, either "dark" or "light".}

\item{width}{The width of the widget as a number or CSS string.}

\item{height}{The height of the widget as a number or CSS string.}

\item{port}{The port for the local web server (which serves local dataset objects to the widget).}

\item{elementId}{An optional element ID.}
}
\description{
The \code{vitessceWidget} function creates a new Vitessce htmlwidget.
A Vitessce widget is defined by a view config which specifies dataset(s),
a layout of views, and linked view connections. A config object can be
passed to the widget using the \code{config} parameter.
}
