% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Component}
\alias{Component}
\title{Component}
\format{
An object of class \code{list} of length 9.
}
\usage{
Component
}
\description{
The \code{Component} list contains an enumeration of
valid string constant values representing components.
}
\keyword{datasets}
