% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FileType}
\alias{FileType}
\title{Enumeration of File Type values}
\format{
An object of class \code{list} of length 12.
}
\usage{
FileType
}
\description{
The \code{FileType} list contains an enumeration of
valid string constant values representing file types
for dataset files.
\itemize{
\item \code{CELLS_JSON}: The \code{cells.json} file type.
\item \code{MOLECULES_JSON}: The \code{molecules.json} file type.
\item \code{NEIGHBORHOODS_JSON}: The \code{neighborhoods.json} file type.
\item \code{RASTER_JSON}: The \code{raster.json} file type.
\item \code{CELL_SETS_JSON}: The \code{cell-sets.json} file type.
\item \code{CLUSTERS_JSON}: The \code{clusters.json} file type.
\item \code{EXPRESSION_MATRIX_ZARR}: The \code{expression-matrix.zarr} file type.
}
}
\examples{
base_url <- "http://localhost:8000/"
vc <- VitessceConfig$new("My config")
dataset <- vc$add_dataset("My dataset")$add_file(
  url = paste0(base_url, "cells.json"),
  data_type = DataType$CELLS,
  file_type = FileType$CELLS_JSON
)
}
\keyword{datasets}
