% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FileType}
\alias{FileType}
\title{File Type}
\format{
An object of class \code{list} of length 6.
}
\usage{
FileType
}
\description{
The \code{FileType} list contains an enumeration of
valid string constant values representing file types
for dataset files.
}
\keyword{datasets}
