% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hconcat}
\alias{hconcat}
\title{Horizontally concatenate views}
\usage{
hconcat(...)
}
\arguments{
\item{...}{A variable number of \code{VitessceConfigView}, \code{VitessceConfigViewHConcat}, or \code{VitessceConfigViewVConcat} objects.}
}
\value{
A \code{VitessceConfigViewHConcat} object.
}
\description{
A helper function to construct a new \code{VitessceConfigViewHConcat} object based on multiple views.
}
\examples{
vc <- VitessceConfig$new(schema_version = "1.0.16", name = "My config")
ds <- vc$add_dataset("My dataset")
spatial <- vc$add_view(ds, Component$SPATIAL)
gene_list <- vc$add_view(ds, Component$FEATURE_LIST)
vc$layout(hconcat(spatial, gene_list))
vc$widget()
}
