% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\docType{class}
\name{SeuratWrapper}
\alias{SeuratWrapper}
\title{SeuratWrapper Class}
\description{
Class representing a local Seurat object in a Vitessce dataset.
}
\details{
Seurat object wrapper class
}
\keyword{internal}
\section{Super class}{
\code{\link[vitessce:AbstractWrapper]{vitessce::AbstractWrapper}} -> \code{SeuratWrapper}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obj}}{The object to wrap.}

\item{\code{cell_set_meta_names}}{The keys in the Seurat object's meta.data
to use for creating cell sets.}

\item{\code{cell_set_meta_name_mappings}}{The keys in the Seurat object's meta.data
to use for cell set names mapped to new names.}

\item{\code{cell_set_meta_score_mappings}}{The keys in the Seurat object's meta.data
to use for cell set names mapped to keys for scores.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SeuratWrapper$new()}}
\item \href{#method-convert_and_save}{\code{SeuratWrapper$convert_and_save()}}
\item \href{#method-create_cells_list}{\code{SeuratWrapper$create_cells_list()}}
\item \href{#method-create_cell_sets_list}{\code{SeuratWrapper$create_cell_sets_list()}}
\item \href{#method-create_expression_matrix_list}{\code{SeuratWrapper$create_expression_matrix_list()}}
\item \href{#method-make_cells_file_def_creator}{\code{SeuratWrapper$make_cells_file_def_creator()}}
\item \href{#method-make_cell_sets_file_def_creator}{\code{SeuratWrapper$make_cell_sets_file_def_creator()}}
\item \href{#method-make_expression_matrix_file_def_creator}{\code{SeuratWrapper$make_expression_matrix_file_def_creator()}}
\item \href{#method-clone}{\code{SeuratWrapper$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="auto_view_config">}\href{../../vitessce/html/AbstractWrapper.html#method-auto_view_config}{\code{vitessce::AbstractWrapper$auto_view_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_file_defs">}\href{../../vitessce/html/AbstractWrapper.html#method-get_file_defs}{\code{vitessce::AbstractWrapper$get_file_defs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_out_dir">}\href{../../vitessce/html/AbstractWrapper.html#method-get_out_dir}{\code{vitessce::AbstractWrapper$get_out_dir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_out_dir_route">}\href{../../vitessce/html/AbstractWrapper.html#method-get_out_dir_route}{\code{vitessce::AbstractWrapper$get_out_dir_route()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_route_str">}\href{../../vitessce/html/AbstractWrapper.html#method-get_route_str}{\code{vitessce::AbstractWrapper$get_route_str()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_routes">}\href{../../vitessce/html/AbstractWrapper.html#method-get_routes}{\code{vitessce::AbstractWrapper$get_routes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vitessce" data-topic="AbstractWrapper" data-id="get_url">}\href{../../vitessce/html/AbstractWrapper.html#method-get_url}{\code{vitessce::AbstractWrapper$get_url()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a wrapper around a Seurat object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$new(
  obj,
  cell_set_meta_names = NA,
  cell_set_meta_score_mappings = NA,
  cell_set_meta_name_mappings = NA,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj}}{The object to wrap.}

\item{\code{cell_set_meta_names}}{An optional list of keys in the object's meta.data
list to use for creating cell sets.}

\item{\code{cell_set_meta_score_mappings}}{If cell_set_meta_names is provided, this list can
also be provided to map between meta.data keys for set annotations
and keys for annotation scores.}

\item{\code{cell_set_meta_name_mappings}}{If cell_set_meta_names is provided, this list can
also be provided to map between meta.data keys and new names to replace
the keys in the interface.}

\item{\code{assay_name}}{The assay name under the assays part of the Seurat object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SeuratWrapper} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_and_save"></a>}}
\if{latex}{\out{\hypertarget{method-convert_and_save}{}}}
\subsection{Method \code{convert_and_save()}}{
Create the JSON output files, web server routes, and file definition creators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$convert_and_save(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_cells_list"></a>}}
\if{latex}{\out{\hypertarget{method-create_cells_list}{}}}
\subsection{Method \code{create_cells_list()}}{
Create a list representing the cells in the Seurat object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$create_cells_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be converted to JSON.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_cell_sets_list"></a>}}
\if{latex}{\out{\hypertarget{method-create_cell_sets_list}{}}}
\subsection{Method \code{create_cell_sets_list()}}{
Create a list representing the cluster assignments in the Seurat object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$create_cell_sets_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be converted to JSON.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_expression_matrix_list"></a>}}
\if{latex}{\out{\hypertarget{method-create_expression_matrix_list}{}}}
\subsection{Method \code{create_expression_matrix_list()}}{
Create a list representing the cluster assignments in the Seurat object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$create_expression_matrix_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list that can be converted to JSON.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_cells_file_def_creator"></a>}}
\if{latex}{\out{\hypertarget{method-make_cells_file_def_creator}{}}}
\subsection{Method \code{make_cells_file_def_creator()}}{
Make the file definition creator function for the cells data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$make_cells_file_def_creator(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A file definition creator function which takes a \code{base_url} parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_cell_sets_file_def_creator"></a>}}
\if{latex}{\out{\hypertarget{method-make_cell_sets_file_def_creator}{}}}
\subsection{Method \code{make_cell_sets_file_def_creator()}}{
Make the file definition creator function for the cell sets data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$make_cell_sets_file_def_creator(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A file definition creator function which takes a \code{base_url} parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_expression_matrix_file_def_creator"></a>}}
\if{latex}{\out{\hypertarget{method-make_expression_matrix_file_def_creator}{}}}
\subsection{Method \code{make_expression_matrix_file_def_creator()}}{
Make the file definition creator function for the expression matrix data type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$make_expression_matrix_file_def_creator(dataset_uid, obj_i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset_uid}}{The ID for this dataset.}

\item{\code{obj_i}}{The index of this data object within the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A file definition creator function which takes a \code{base_url} parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeuratWrapper$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
