% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vconcat}
\alias{vconcat}
\title{Vertically concatenate views}
\usage{
vconcat(...)
}
\arguments{
\item{...}{A variable number of \code{VitessceConfigView}, \code{VitessceConfigViewHConcat}, or \code{VitessceConfigViewVConcat} objects.}
}
\value{
A \code{VitessceConfigViewVConcat} object.
}
\description{
A helper function to construct a new \code{VitessceConfigViewVConcat} object based on multiple views.
}
