% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{VitessceConfigServerRangeRoute}
\alias{VitessceConfigServerRangeRoute}
\title{Class representing a local web server route for a file which needs to support range requests.}
\description{
Class representing a local web server route for a file which needs to support range requests.

Class representing a local web server route for a file which needs to support range requests.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{The path on which the web server should respond to requests using this callback.}

\item{\code{file_path}}{The file to serve.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VitessceConfigServerRangeRoute$new()}}
\item \href{#method-create_handlers}{\code{VitessceConfigServerRangeRoute$create_handlers()}}
\item \href{#method-clone}{\code{VitessceConfigServerRangeRoute$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new server route wrapper object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServerRangeRoute$new(path, file_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The route path.}

\item{\code{file_path}}{The file to serve on this route.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VitessceConfigServerRangeRoute} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_handlers"></a>}}
\if{latex}{\out{\hypertarget{method-create_handlers}{}}}
\subsection{Method \code{create_handlers()}}{
Add handler functions to the Plumber server object to respond on this route.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServerRangeRoute$create_handlers(pr_server)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pr_server}}{The server instance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The modified server instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VitessceConfigServerRangeRoute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
